Move VecIndex.h/VecIndexFactory.cpp/VecIndexFactory.h to directory index
