+++ knowhere/index/vector_index/IndexAnnoy.h
@@ -17,7 +17,7 @@
 #include "annoy/src/kissrandom.h"
 
 #include "knowhere/common/Exception.h"
-#include "knowhere/index/vector_index/VecIndex.h"
+#include "knowhere/index/VecIndex.h"
 
 namespace knowhere {
