+++ knowhere/index/vector_index/IndexBinaryIVF.h
@@ -18,8 +18,8 @@
 #include <faiss/IndexIVF.h>
 
 #include "knowhere/common/Exception.h"
+#include "knowhere/index/VecIndex.h"
 #include "knowhere/index/vector_index/FaissBaseBinaryIndex.h"
-#include "knowhere/index/vector_index/VecIndex.h"
 
 namespace knowhere {
