+++ knowhere/index/vector_index/IndexIDMAP.h
@@ -15,8 +15,8 @@
 #include <utility>
 #include <vector>
 
+#include "knowhere/index/VecIndex.h"
 #include "knowhere/index/vector_index/FaissBaseIndex.h"
-#include "knowhere/index/vector_index/VecIndex.h"
 
 namespace knowhere {
