+++ knowhere/index/vector_index/IndexIVF.h
@@ -18,8 +18,8 @@
 #include <faiss/IndexIVF.h>
 
 #include "knowhere/common/Typedef.h"
+#include "knowhere/index/VecIndex.h"
 #include "knowhere/index/vector_index/FaissBaseIndex.h"
-#include "knowhere/index/vector_index/VecIndex.h"
 
 namespace knowhere {
