+++ knowhere/index/vector_offset_index/IndexIVF_NM.h
@@ -19,7 +19,7 @@
 #include <faiss/IndexIVF.h>
 
 #include "knowhere/common/Typedef.h"
-#include "knowhere/index/vector_index/VecIndex.h"
+#include "knowhere/index/VecIndex.h"
 #include "knowhere/index/vector_offset_index/OffsetBaseIndex.h"
 
 namespace knowhere {
