+++ unittest/test_ivf.cpp
@@ -19,7 +19,7 @@
 
 #include "knowhere/common/Exception.h"
 #include "knowhere/index/IndexType.h"
-#include "knowhere/index/vector_index/VecIndexFactory.h"
+#include "knowhere/index/VecIndexFactory.h"
 #include "knowhere/index/vector_index/adapter/VectorAdapter.h"
 
 #ifdef KNOWHERE_GPU_VERSION
