+++ unittest/test_ivf_nm.cpp
@@ -20,7 +20,7 @@
 
 #include "knowhere/common/Exception.h"
 #include "knowhere/index/IndexType.h"
-#include "knowhere/index/vector_index/VecIndexFactory.h"
+#include "knowhere/index/VecIndexFactory.h"
 #include "knowhere/index/vector_index/adapter/VectorAdapter.h"
 
 #ifdef KNOWHERE_GPU_VERSION
